requirements.txt
pandas>=2.2.0
numpy
joblib
scikit-learn
requests>=2.31.0
python-dotenv>=1.0.0
tqdm
customtkinter
Pillow
requests-cache
matplotlib